+++ lib/app_ops_type.dart
@@ -2,4 +2,5 @@ enum AppOpsType {
   usageStats,
   systemAlertWindow,
   pictureInPicture,
+  writeSettings,
 }