+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.2
 
+* feat: Add Write Settings
 * feat: Add Picture In Picture
 
 ## 0.0.1